Add shaders
This commit adds shaders :)
To make a new shader you make a folder in shaders/ along with a fragment
shader, a vertex shader and a uniforms file. This will then be added to
shaders.js in gen and will be available as a shader with the same name
as your folder. For now you will need to have all three shaders there,
but this will be improved