Add default fallback to shadergen
This commits adds a fallback if you don't want to have every part of a
shader ín your shader. You will have to oblige to the standard imposed
by me for shaders.

You will need to have the tDiffuse in your uniforms file.
The vertex shader needs to pass vUv (uv) on to the fragment shader.

Other than this it should be good :)