Make test-project working fodlder
Now the test-project folder is a working directory. This means that the
project you are currently working on always lies inside test-project. If
you want to change what you are working on, change the contents of
test-project. This will make it easier to work on both nin and another
project at the same time.

All example code has been moved into an example folder. This folder is
intended to be a catalog with effects and so on. If you want to add an
effect we might use, please add it to the exampple folder.