Move nin backend to own folder
Since the nin backend lives in its own folder, it becomes more
difficult to maintain links to hardcoded paths such as "test-project".
This commit changes therefore also nin to use configurable project
paths.

The backend also begins to serve the compiled version of the frontend
with this commit. The advantage of this is that if you add the `nin`
executable to your $PATH, you can begin to use nin as it was intended
from day one: simply write run `nin` directly from the project folder.