Change camerapaths format
This changes the semantics of the connection between a camerapath and a
layer slightly.
The old format matched camerapaths in camerapaths.json with
corresponding layers in layers.json by their index in the list. This
made it difficult to actually keep a camerapath connected to the proper
layer.

A full rework of the system (diffent camerapaths in different files or
something similar) might be best, but that's too much work to be
completed before solskogen 2015.

The old format meant that a camerapath was linked to an instance of the
layer, i.e. two different instances of the layer could (and would) have
different camerapaths. With the new system, the camerapath is instead
linked to the 'type' of the layer, as `camerapaths.json` is no longer a
list, but a dictionary where the layer type is the key for each
dictionary.
This should make it easier to manage the relation between layers and
camerapaths, and I doubt the possibility of using different camerapaths
for different instances of the same layer will be missed.