Split cameracontroller and create pathcontroller
Moving the most generic parts of the cameracontroller to a
pathcontroller will allow for using the cameracontroller system for
animating more than the camera alone.

A pathcontroller may be instantiated with a javascript object similar to
the ones used by each property of the camera controller.

The pathcontroller knows nothing about how these objects are loaded,
which means they may either be stored in an external JSON file, or
simply as an inline js object.

Typical usage will be:

```
var raw_path = [
  {
    'startFrame': 0,
    'endFrame': 200,
    'points': [
      [0, 0, 0],
      [0, 100, 0]
      [100, 100, 0]
    ]
  },
  ...
];
this.fancyObjectPath = new PathController(raw_path, '3D');
```

Then, in the update loop of your layer, call
`this.fancyObjectPath.get3DPoint(relativeFrame);` to obtain the current
position for your object.