Change music to be webaudioapi-based
The music is no longer played by a HTML5 audio element, but rather by
the webaudioapi. This is done to allow more interesting music
integration down the line. As an implementation result, demo.music is
now an interface object that manipulates webaudioapi playback, rather
than an audio object. The interface has been changed somewhat, which is
reflected in the usage of demo.music in this commit.

The music is loaded and played back as an AudioBuffer. This is a method
recommended for "short to medium-length audio", as it loads the entire
buffer into memory. An alternative to AudioBuffer would be to use a
traditional audio element and run it through the webaudio pipeline using
a MediaElementSourceNode, but this is not done since it allegedly has
flaky support across browsers, and it limits some interesting webaudio
opportunities that would otherwise be available to us, such as a
non-pitch-correcting playbackRate modifier.

Since AudioBufferSourceNodes are designed as a cheap, one-off node for
playing short sound bytes (think sound effects), there are no playback
controls, so these have been implemented explicitly in this commit.