Fix loadAjax bug where old-format calls would die
The loadAjax signature was changed from loadAjax(path, callback) to
loadAjax(path, options, callback), and to make things backwards
compatible, the options parameter in the middle is optional, letting old
calls still work. However, there was a bug in the implementation of this
backwards compatibility. This bug is fixed in this commit.