Set clear flag on render pass to false
The clear flag needs to be false to allow transparent render passes on
top of eachother. It was originally set to false in nin, but updating
three.js resets it to true, since that's the default in the official
three.js version.

This commit is a copy of 7ec1acca3327e4de25a68541d70b9a477209a927, but
needs to be re-applied because the official lib from three.js sets clear
to true.