Switch Grunt to webpack
This is instead of migrating to a anew major version of grunt and trying
to update all of its deprecated plugins.
In the process, all bower dependencies are also moved over to npm so now
we only need one package manager.

In the process, all files needed to be modularized.
When doing that, all controllers and services were migrated to ES6 class
syntax.
For ES6 class syntax compatibility, Angular's controllerAs pattern is also employed.