Publish nin on NPM
The way this works is that when we publish nin on NPM we go to the
backend folder and type `make publish`. The account is called ninjadev
who has this package. A user of nin can then type `npm install -g
ninjadev-nin` to install the backage and get `nin` as a cli.

When publishing on NPM the frontend is compiled and copied to the
backend folder, the dasBoot folder is also copied into the backend
folder. A sed script changes the serve.js file to serve the files in the
backend folder.