Lint remaining backend files
The lint command in package.json is updated so that the eslint runs with
backend as cwd. This ensures that it loads the correct .eslintignore

Templates in the generate folder are ignored since they're not strictly
part of the backend, but templates for user projects.

Also, one lint error remains, but it's there because the feature is
still broken so I figured it serves as a nice reminder to fix the
feature.