Replace closure compiler with js implementation
The Google Closure Compiler now has an official js port, so we can use
that instead. The upside for us is that we no longer have a Java
dependency, and we can have a cleaner-looking compile implementation
that doesn't rely on external executables.

Of course, compilation is currently broken in master since the switch to
nodes instead of layers, so compiling anything on this commit won't
technically work. However, this should "totally work" when we fix the
other stuff.