Add super class for THREE-based nodes
This simplifies using the cameracontroller by not requiring the
boilerplate of a separate cameracontroller node, and instead handling
the setup in the THREENode superclass.
This also lays the foundation for a flyaround mode that will be
made available for all nodes extending NIN.THREENode.

To update code currently using the cameracontroller, change graph.json
from:

```
[
...
{
  "id": "node",
  "type": node",
  "connected": {
    "camera": "nodecamera.camera"
  }
},
{
  "id": "nodecamera",
  "type": NIN.CameraNode",
  "options": {
    "path": "res/node.camera.json"
  }
},
...
]
```

to:
```
[
...
{
  "id": "node",
  "type": node",
  "options": {
    "camera": "res/node.camera.json"
  }
},
...
]
```