Add better font support
Fonts can now be added to the res/ folder, and they will be
automagically be made available as a `@font-face` with the same name as
the font file (without suffix). As an example: `res/myfont.otf` will be
available for a 2D canvas as `ctx.font = '10pt myfont';`. Only .otf is
supported for now.

Before this commit, the typical way of adding fonts was to manually
convert otf to base64 offline, load that with Loader.loadAjax, and take
care to create a style tag with the font definition somewhere in the
demo.