Add first steps towards an ui-editable graph
Now graph edges can be added and removed in a crude first-step
implementation of an editable ui. The plan is to slowly iterate towards
a slick solution over multiple pull requests.

This pull requests disables live watch reloads on res/graph.json. The
idea is that it is loaded from the backend only once, and from then on
the frontend is responsible for editing it, sending graph change events
to the backend for persistance.