Upgrade chokidar and webpack for node 10 compat
Both chokidar and webpack depended on old versions of fsevents, which is
used for file watching on macOS. That version of fsevents did not come
with pre-build binaries for node 10 which resulted in install issues on
my machine. These new versions of chokidar and webpack depend on the
latest version of fsevents which has binaries for node 10 pre-built,
without losing support for older node versions (even down to 4.x).

Luckily, upgrading from webpack 2.x to 4.x is trivial compared to
upgrading from webpack 1.x to 2.x \o/