Fix beat bean initing
After upgrading webpack, the generated code in dasboot.js changed for
the cases where we do:
```
const {functionA, functionB} = require('./module');
window.functionA = functionA;
window.functionB = functionB;
```

When the closure compiler sees this during compilation of the demo, it
derps and doesn't recognize that we might actually use the local
variable `functionA` inside bootstrap.js directly, and removes the
variable.

Updating the code in this file to read the function from the window
scope explicitly works as a hack to make compilation work again.

In the future, we may fix this properly by removing everything from
window scope and using modules fully.