Use travis documented way of reading env variables
https://docs.travis-ci.com/user/conditions-v1

Better than guessing!

Also, this moves tagging of the commit, because I could not figure out
how to use an environment variable as part of the tag in the npm
provider template. This might not work, so the next things to test is to
also add a provider block for GitHub releases that will also mark the
publish as a release in GitHub and set the tag at the same time.