Fix syntax error in .travis.yml
I think I _finally_ understood the docs on this feature now:

> condition: deploy when a single bash condition evaluates to true. This
must be a string value, and is equivalent to if [[ <condition> ]]; then
<deploy>; fi. For example, $CC = gcc.

The docs were inconsistent on whether or not the whole thing was wrapped
in strings. I think now, that if it's all in quotes it will be executed
at runtime as a bash if statement. Otherwise, it is evaluated by travis
and the _result_ is put as the condition in bash. So for instance,
travis can evaluate something to "true" and then put the string "true"
in if [[ true ]]; then

Hopefulle this should do the trick, and for good measure I included a
guard to ensure this only runs on master, now that I might have
overridden the conditions for when deploy runs.